requirements.txt
absl-py==2.0.0
aiofiles @ file:///croot/aiofiles_1683773582346/work
aiosqlite @ file:///croot/aiosqlite_1683773899903/work
alembic==1.10.3
anyio @ file:///tmp/build/80754af9/anyio_1644481695334/work/dist
argon2-cffi @ file:///home/builder/ci_310/argon2-cffi_1641902187184/work
asttokens @ file:///opt/conda/conda-bld/asttokens_1646925590279/work
async-generator==1.10
attrs==22.2.0
Babel @ file:///croot/babel_1671781930836/work
backcall @ file:///home/ktietz/src/ci/backcall_1611930011877/work
beautifulsoup4==4.12.2
bleach==6.0.0
brotlipy==0.7.0
cachetools==5.3.2
certifi @ file:///croot/certifi_1690232220950/work/certifi
certipy==0.1.3
cffi==1.15.1
charset-normalizer @ file:///tmp/build/80754af9/charset-normalizer_1630003229654/work
comm @ file:///croot/comm_1671231121260/work
contourpy @ file:///opt/conda/conda-bld/contourpy_1663827406301/work
cryptography==40.0.1
cycler @ file:///tmp/build/80754af9/cycler_1637851556182/work
debugpy @ file:///croot/debugpy_1690905042057/work
decorator @ file:///opt/conda/conda-bld/decorator_1643638310831/work
defusedxml==0.7.1
entrypoints @ file:///tmp/build/80754af9/entrypoints_1649908313000/work
exceptiongroup @ file:///croot/exceptiongroup_1668714342571/work
executing @ file:///opt/conda/conda-bld/executing_1646925071911/work
fastjsonschema==2.16.3
flit_core @ file:///opt/conda/conda-bld/flit-core_1644941570762/work/source/flit_core
fonttools==4.25.0
google-auth==2.23.4
google-auth-oauthlib==1.1.0
greenlet==2.0.2
grpcio==1.59.2
idna @ file:///croot/idna_1666125576474/work
ipykernel @ file:///croot/ipykernel_1691121631942/work
ipython @ file:///croot/ipython_1694181358621/work
ipython-genutils @ file:///tmp/build/80754af9/ipython_genutils_1606773439826/work
ipywidgets @ file:///croot/ipywidgets_1679394798311/work
jedi @ file:///tmp/build/80754af9/jedi_1644315229345/work
Jinja2==3.1.2
json5 @ file:///tmp/build/80754af9/json5_1624432770122/work
jsonschema==4.17.3
jsonschema-specifications @ file:///croot/jsonschema-specifications_1699032386549/work
jupyter @ file:///tmp/abs_33h4eoipez/croots/recipe/jupyter_1659349046347/work
jupyter-console @ file:///croot/jupyter_console_1679999630278/work
jupyter-events @ file:///croot/jupyter_events_1699282461638/work
jupyter-server @ file:///croot/jupyter_server_1671707632269/work
jupyter-telemetry==0.1.0
jupyter-ydoc @ file:///croot/jupyter_ydoc_1683747223142/work
jupyter_client==8.1.0
jupyter_core==5.3.0
jupyter_server_fileid @ file:///croot/jupyter_server_fileid_1684273577568/work
jupyter_server_ydoc @ file:///croot/jupyter_server_ydoc_1686767404829/work
jupyterhub==3.1.1
jupyterlab @ file:///croot/jupyterlab_1686179668131/work
jupyterlab-pygments==0.2.2
jupyterlab-widgets @ file:///croot/jupyterlab_widgets_1679055282532/work
jupyterlab_server @ file:///croot/jupyterlab_server_1680792501448/work
kiwisolver @ file:///opt/conda/conda-bld/kiwisolver_1653292039266/work
Mako==1.2.4
Markdown==3.5.1
MarkupSafe==2.1.2
matplotlib @ file:///croot/matplotlib-suite_1670466153205/work
matplotlib-inline @ file:///opt/conda/conda-bld/matplotlib-inline_1662014470464/work
mistune==2.0.5
mkl-fft==1.3.1
mkl-random @ file:///home/builder/ci_310/mkl_random_1641843545607/work
mkl-service==2.4.0
munkres==1.1.4
nb-conda-kernels @ file:///home/builder/ci_310/nb_conda_kernels_1640815331881/work
nbclassic @ file:///croot/nbclassic_1699542793266/work
nbclient==0.7.3
nbconvert==7.3.1
nbformat==5.8.0
nest-asyncio @ file:///croot/nest-asyncio_1672387112409/work
notebook @ file:///croot/notebook_1690984815942/work
notebook_shim @ file:///croot/notebook-shim_1699455894279/work
numpy @ file:///croot/numpy_and_numpy_base_1668593735768/work
oauthlib==3.2.2
opencv-python==4.8.1.78
packaging==23.1
pamela==1.0.0
pandocfilters==1.5.0
parso @ file:///opt/conda/conda-bld/parso_1641458642106/work
pexpect @ file:///tmp/build/80754af9/pexpect_1605563209008/work
pickleshare @ file:///tmp/build/80754af9/pickleshare_1606932040724/work
Pillow==9.2.0
platformdirs==3.2.0
ply==3.11
prometheus-client==0.16.0
prompt-toolkit @ file:///croot/prompt-toolkit_1672387306916/work
protobuf==4.23.4
psutil @ file:///opt/conda/conda-bld/psutil_1656431268089/work
ptyprocess @ file:///tmp/build/80754af9/ptyprocess_1609355006118/work/dist/ptyprocess-0.7.0-py2.py3-none-any.whl
pure-eval @ file:///opt/conda/conda-bld/pure_eval_1646925070566/work
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
Pygments==2.15.0
pyOpenSSL==23.1.1
pyparsing @ file:///opt/conda/conda-bld/pyparsing_1661452539315/work
PyQt5-sip==12.11.0
pyrsistent==0.19.3
PySocks @ file:///home/builder/ci_310/pysocks_1640793678128/work
python-dateutil==2.8.2
python-json-logger==2.0.7
pytz @ file:///croot/pytz_1695131579487/work
PyYAML @ file:///croot/pyyaml_1698096049011/work
pyzmq==25.0.2
qtconsole @ file:///croot/qtconsole_1681394213385/work
QtPy @ file:///opt/conda/conda-bld/qtpy_1662014892439/work
referencing @ file:///croot/referencing_1699012038513/work
requests @ file:///opt/conda/conda-bld/requests_1657734628632/work
requests-oauthlib==1.3.1
rfc3339-validator @ file:///croot/rfc3339-validator_1683077044675/work
rfc3986-validator @ file:///croot/rfc3986-validator_1683058983515/work
rpds-py @ file:///croot/rpds-py_1698945930462/work
rsa==4.9
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
Send2Trash @ file:///croot/send2trash_1699371139552/work
sip @ file:///tmp/abs_44cd77b_pu/croots/recipe/sip_1659012365470/work
six @ file:///tmp/build/80754af9/six_1644875935023/work
sniffio @ file:///home/builder/ci_310/sniffio_1640794799774/work
soupsieve==2.4
SQLAlchemy==2.0.9
stack-data @ file:///opt/conda/conda-bld/stack_data_1646927590127/work
tensorboard==2.15.1
tensorboard-data-server==0.7.2
terminado @ file:///croot/terminado_1671751832461/work
tinycss2==1.2.1
toml @ file:///tmp/build/80754af9/toml_1616166611790/work
tomli @ file:///opt/conda/conda-bld/tomli_1657175507142/work
torch==1.13.0
torchaudio==0.13.0
torchvision==0.14.0
tornado==6.2
tqdm==4.66.1
traitlets==5.9.0
typing_extensions==4.5.0
urllib3 @ file:///croot/urllib3_1670526988650/work
wcwidth @ file:///Users/ktietz/demo/mc3/conda-bld/wcwidth_1629357192024/work
webencodings==0.5.1
websocket-client @ file:///home/builder/ci_310/websocket-client_1640795866898/work
Werkzeug==3.0.1
widgetsnbextension @ file:///croot/widgetsnbextension_1679313860248/work
y-py @ file:///croot/y-py_1683555143488/work
ypy-websocket @ file:///croot/ypy-websocket_1684171737040/work

numpy~=1.23.4